% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-autouse_crosswalk_simple.R
\docType{data}
\name{autouse_crosswalk_simple}
\alias{autouse_crosswalk_simple}
\title{Simplified AU to Use crosswalk}
\format{
A data frame with N rows and M variables:
\describe{
  \item{col1}{…}
  \item{col2}{…}
  …
}
}
\source{
Your source here
}
\usage{
autouse_crosswalk_simple
}
\description{
A simplified version of the `autouse_crosswalk_raw` dataset, containing only select columns.
}
\keyword{datasets}
