% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{get_df_ATTAINS_orgs}
\alias{get_df_ATTAINS_orgs}
\title{Get valid ATTAINS organization names}
\usage{
get_df_ATTAINS_orgs()
}
\value{
A character vector of organization names
}
\description{
This helper fetches domain values for org_id from rExpertQuery
and returns the name column for use in UI components.
}
