% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mltoau_crosswalk_simple.R
\docType{data}
\name{mltoau_crosswalk_simple}
\alias{mltoau_crosswalk_simple}
\title{Simplified ML to AU crosswalk}
\format{
A data frame with N rows and M variables:
\describe{
  \item{col1}{…}
  \item{col2}{…}
  …
}
}
\source{
Your source here
}
\usage{
mltoau_crosswalk_simple
}
\description{
A simplified version of the `mltoau_crosswalk_raw` dataset, containing only select columns.
}
\keyword{datasets}
