% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imports_globalvars.R
\docType{package}
\name{TADAShinyJoinToAU-package}
\alias{TADAShinyJoinToAU}
\alias{TADAShinyJoinToAU-package}
\title{TADAShinyJoinToAU: TADAShinyJoinToAU: Join Monitoring Locations to AUs}
\description{
The overall goal of the `TADAShinyJoinToAU` shiny app is to provide a second module to the `TADAShiny` shiny app where users can join monitoring locations to their respective assessment units (AUs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USEPA/TADAShinyJoinToAU}
  \item Report bugs at \url{https://github.com/USEPA/TADAShinyJoinToAU/issues}
}

}
\author{
\strong{Maintainer}: Sheila Saia \email{sheila.saia@tetratech.com} (\href{https://orcid.org/0000-0002-4943-0306}{ORCID})

Authors:
\itemize{
  \item Sheila Saia \email{sheila.saia@tetratech.com} (\href{https://orcid.org/0000-0002-4943-0306}{ORCID})
  \item Cristina Mullin \email{mywaterway@epa.gov} (\href{https://orcid.org/0000-0002-0615-6087}{ORCID})
  \item Ben Block \email{ben.block@tetratech.com} (\href{https://orcid.org/0000-0003-1427-4363}{ORCID})
  \item Yu-Chen Wang
}

}
\keyword{internal}
